python: automatically restart python binary with ld path
There is a very tricky issue in the dynamic library loading.
The root cause is that e.g. `OpenBLAS_jll` runs a dynamic call `dlopen("libgfortran.so")`.
Thus, when the `poseest` lib is loaded at runtime by the python library,
we need to dynamically link. (This is as opposed to building e.g.
`mainc`, where the linking can happen at compile time(?)).

During dynamic linking, apparently a "correct" implementation of the
linker ignores the RPATH, so that libgfortran can't be found anymore.
Some distros such as ubuntu have some hacks built in that make this work
regardless, but it would be bad to rely on those, and it doesn't work
e.g. on alpine linux (which we use to try to build the musl libc based version).

One could also think that it would be sufficient to just set the
LD_LIBRARY_PATH envvar at runtime in the python script. However,
apparently you can't modify the library path after the python process
has spawned.

Thus, we make this crazy modification. We check if the library path is
available in the LD_LIBRARY_PATH variable. If not, we add it to the
environment and then relaunch the python executable (!). This will
trigger a reload of all the code (e.g., other imported packages), but
with a fixed path.