Restore juliac by removing `merge!` during init (units), `vcat` better, and no C init
I don't remember what we need the C init function for but we shouldn't
need it. The unitful thing is just annoying when trying to use units
from python... we'll need to find a solution for that, but perhaps I can
move this code into the extension. Finally, the vcat shouldn't change
anything, although we do get a warning now about static arrays, but that
should be fine.